- plus de problèmes avec les arrondis des taille des fontes dans copier/coller, les tailles des fontes sont maintenant des double
git-svn-id: https://svn.opac.ch/svn/cr/branches/vs2005@5888 787ff09c-d2b5-7842-b444-8dd0a418631d

+++ Epsitec.Cresus/Common.Text.Exchange/HtmlText.cs
@@ -195,7 +195,7 @@ namespace Epsitec.Common.Text.Exchange
 
 		public void SetFontSize(double fontSize)
 		{
-			this.fontSize = (int) (fontSize / /* Epsitec.Common.Document.Modifier.FontSizeScale*/ FontSizeFactor);
+			this.fontSize = fontSize / /* Epsitec.Common.Document.Modifier.FontSizeScale*/ FontSizeFactor;
 		}
 
 		static private int[] htmlfontsizes =
